pyproject.toml
[project]
name = "hush"
version = "0.1.0"
description = "Smart notifications for Home Assistant - the notification system HA should have shipped with"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "mossipcams" }]
keywords = ["homeassistant", "notifications", "smart-home"]

[project.urls]
Homepage = "https://github.com/mossipcams/hush"
Repository = "https://github.com/mossipcams/hush"
Documentation = "https://github.com/mossipcams/hush"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["custom_components.hush"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "homeassistant.*",
    "aiosqlite.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["custom_components/hush"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-homeassistant-custom-component>=0.13.0",
    "pytest-cov>=6.0.0",
]
